Updated action() in BreweryAction and added it to FieldController. Updated csv
